# init: workers channel batches