CPU workers to produce label groups in batches, in attempt to reduce channel ops overhead.

## Benchmarks

### INIT 2GB | 6 in-file parallelism
 before: 2m7s
 after: 1m58s

time % before | time % after | function 
-- | -- | -- 
24 | 39 | github.com/spacemeshos/sha256-simd.blockAvx2 (cpu)
6 | 8 | syscall.Syscall (disk write)
14 | 12 | runtime.pthread_cond_signal (chan ops)
26 | 16 | runtime.pthread_cond_wait (chan ops)
8 | 1 | runtime.usleep 

### INIT 2GB | no parallelism
before: 10m7s
after: 7m43s
**no channels: 8m32s**


time % before | time % after | time % no-channels | function 
-- | -- | -- | --
0.5 | 6 | 11 |  github.com/spacemeshos/sha256-simd.blockAvx2 (cpu)
3 | 5 | 75 | syscall.Syscall (disk write)
59 | 73 | 2 | runtime.pthread_cond_signal (chan ops)
23 | 9 | 2 | runtime.pthread_cond_wait (chan ops)
10 | 1 | 0 | runtime.usleep